Make schedule more concrete
Make sure we know what lectures are on what days. This also lets us know that,
despite all the holidays, we have one flex day if we fall behind.

This also flips the order of VCS and Linux, so that the students are
more comfortable with the command line before learning Git.