Push most homeworks back by one lecture
We planned the Tuesday homeworks with the (wrong) assumption that class
would begin on a Tuesday. Since it actually began on a Thursday, we
should move the release/due dates to Thursdays as well to better match
the lecture content.

Note that the homeworks for the final module are still due when they
were before, because we want to keep one lecture of buffer at the end in
case we need to extend one of the later assignments.