Assignment 1: explicitly forbid /dev/null only
We may not have mentioned /dev/null in lecture by the time students read
the assignment, so let's explicitly tell them that it's coming. This
also means we won't accidentally punish students who do the assignment
early if we mention something else from /dev/ when answering a question
in lecture.