Assignment 4: fix relative link
It appears that Jekyll does some processing on Markdown relative links
to ensure they're properly resolved even when pretty URLs are on.
However, this processing doesn't seem to catch links that are split over
two lines of Markdown source. Put the link on a single line so it'll
resolve properly on the real site.